+++ src/test/canonical_tests.cpp
@@ -6,12 +6,11 @@
 // Unit tests for canonical signatures
 //
 
-
-
-#include "script.h"
-#include "util.h"
 #include "data/sig_noncanonical.json.h"
 #include "data/sig_canonical.json.h"
+#include "random.h"
+#include "script.h"
+#include "util.h"
 
 #include <boost/foreach.hpp>
 #include <boost/test/unit_test.hpp>
@@ -20,7 +19,6 @@
 
 using namespace std;
 using namespace json_spirit;
-
 
 // In script_tests.cpp
 extern Array read_json(const std::string& jsondata);
