+++ src/test/test_bitcoin.cpp
@@ -4,9 +4,8 @@
 
 #define BOOST_TEST_MODULE Bitcoin Test Suite
 
-
-
 #include "main.h"
+#include "random.h"
 #include "txdb.h"
 #include "ui_interface.h"
 #include "util.h"
@@ -89,4 +88,3 @@ bool ShutdownRequested()
 {
   return false;
 }
-
