+++ src/test/util_tests.cpp
@@ -4,6 +4,7 @@
 
 #include "util.h"
 
+#include "random.h"
 #include "sync.h"
 
 #include <stdint.h>
